pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "event_api_service"
version = "0.1.0"
authors = [
  { name="Event API Team", email="contact@example.com" },
]
description = "Unified Event API Service - Aggregate events from Eventbrite, Meetup, Ticketmaster, and more"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Flask",
    "Flask-Limiter",
    "python-dotenv",
    "requests",
    "geopy"
]

[project.urls]
"Homepage" = "https://github.com/example/event_api"
